Workaround HEAD responses with undefined lengths
While Cowboy's strict and always requires to put a content-length in a
response to a HEAD request, the clients Heroku proxies for do not always
end up doing so, and leave the content-length value as missing.

In usual circumstances, this would mean that the connection is
close-delimited, and this is how we treat it (we can't know if it's an
'unknown' value that needs to be streamed, or an omission, and both
happen in the wild).

The issue is that because HEAD responses are often used as an
optimization, closing the connection on these (because we faithfully
replicate the expected 'connection:close' header) ends up being
inefficient in many cases.

Because of this, we use a streaming function with an unknown length that
shortcircuits the creation of content-length headers -- which the
streaming function is meant for -- The issue is that in some
circumstances (HEAD requests) we may want to override the 'close' value
on a case-by-case basis.

This patch lets us do this.