improve token capitalization performance
rewrite it to more heavily uses bifs and to keep the intermediate state
in an iolist rather than a binary.  for whatever nonsense beam reasons
this is faster.  could have something to do with tiny binaries and
refcounting but I can't even speculate.  This solution approaches the
speed of just lowercasing the entire binary in a binary comprehension,
so I don't imagine we can get much faster.